GROK-15018: Chem: RGroup analysis improvements (highlight r groups, replace latest analysis)

+++ packages/Chem/src/analysis/r-group-analysis.ts
@@ -6,10 +6,12 @@ import {getRdKitModule} from '../package';
 import {getMCS} from '../utils/most-common-subs';
 import {RDMol} from '@datagrok-libraries/chem-meta/src/rdkit-api';
 import { _convertMolNotation } from '../utils/convert-notation-utils';
-import { SCAFFOLD_COL } from '../constants';
+import { SCAFFOLD_COL, SUBSTRUCT_COL } from '../constants';
 import {MolfileHandler} from '@datagrok-libraries/chem-meta/src/parsing-utils/molfile-handler';
+import { getUncommonAtomsAndBonds } from '../utils/chem-common-rdkit';
 
-const latestAnalysisCols: string[] = [];
+
+let latestAnalysisCols: string[] = [];
 let latestTrellisPlot: DG.Viewer | null = null;
 
 export function convertToRDKit(smiles: string): string {
@@ -32,10 +34,11 @@ export function rGroupAnalysis(col: DG.Column): void {
   const visualAnalysisCheck = ui.boolInput('Visual analysis', true);
   const exactAtomsCheck = ui.boolInput('MCS exact atoms', true);
   const exactBondsCheck = ui.boolInput('MCS exact bonds', true);
+  const replaceLatest = ui.boolInput('Replace latest', true);
   const undoLatestAnalysis = ui.icons.undo(() => {
-    latestTrellisPlot?.close();
-    latestAnalysisCols.forEach((colName: string) => col.dataFrame.columns.remove(colName));
+    removeLatestAnalysis(col);
   }, 'Undo latest analysis');
+  undoLatestAnalysis.classList.add('chem-rgroup-undo-icon');
 
   const molColNames = col.dataFrame.columns.bySemTypeAll(DG.SEMTYPE.MOLECULE).map((c) => c.name);
   const columnInput = ui.choiceInput('Molecules', col.name, molColNames);
@@ -73,22 +76,23 @@ export function rGroupAnalysis(col: DG.Column): void {
       ]),
       columnInput,
       columnPrefixInput,
-      visualAnalysisCheck.root,
-      undoLatestAnalysis
+      ui.divH([visualAnalysisCheck.root, replaceLatest.root, undoLatestAnalysis])
     ]))
     .onOK(async () => {
+      if (replaceLatest.value)
+        removeLatestAnalysis(col);
       const getPrefixIdx = (colPrefix: string) => {
         let prefixIdx = 0;
         col = col.dataFrame.columns.byName(columnInput.value!);
         const re = new RegExp(`^${colPrefix}$`, 'i');
         if (col.dataFrame.columns.names().filter(((it) => it.match(re))).length) {
-          prefixIdx++;  
+          prefixIdx++;
           const maxPrefixIdx = 100;
           for (let i = 0; i < maxPrefixIdx; i++) {
             const reIdx = new RegExp(`^${colPrefix}_${prefixIdx}$`, 'i');
-            if(!col.dataFrame.columns.names().filter(((it) => it.match(reIdx))).length)
+            if (!col.dataFrame.columns.names().filter(((it) => it.match(reIdx))).length)
               break;
-            prefixIdx++;       
+            prefixIdx++;
           }
           if (prefixIdx - 1 === maxPrefixIdx) {
             grok.shell.error('Table contains columns named \'R[number]\', please change column prefix');
@@ -111,24 +115,30 @@ export function rGroupAnalysis(col: DG.Column): void {
       let progressBar;
       try {
         const onlyMatchAtRGroups =
-          !!MolfileHandler.getInstance(core).atomTypes.filter((it) => it.startsWith('R')).length && 
+          !!MolfileHandler.getInstance(core).atomTypes.filter((it) => it.startsWith('R')).length &&
           core.includes('M  RGP');
         progressBar = DG.TaskBarProgressIndicator.create(`RGroup analysis running...`);
         const res = await findRGroupsWithCore(col.name, col.dataFrame, core, onlyMatchAtRGroups);
         const module = getRdKitModule();
         if (res.rowCount) {
+          //unmatched are those items for which all R group cols are empty
+          const unmatchedItems = new Uint8Array(res.rowCount).fill(0);
           for (const resCol of res.columns) {
             const molsArray = new Array<string>(resCol.length);
             for (let i = 0; i < resCol.length; i++) {
               const molStr = resCol.get(i);
-              let mol: RDMol | null = null;
-              try {
-                mol = module.get_mol(molStr);
-                molsArray[i] = mol.get_molblock().replace('ISO', 'RGP');
-              } catch (e) {
-                //do nothing here, molsArray[i] is empty for invalid molecules
-              } finally {
-                mol?.delete();
+              if (resCol.name !== 'Core' && !molStr) {
+                unmatchedItems[i] += 1;
+              } else {
+                let mol: RDMol | null = null;
+                try {
+                  mol = module.get_mol(molStr);
+                  molsArray[i] = mol.get_molblock().replace('ISO', 'RGP');
+                } catch (e) {
+                  //do nothing here, molsArray[i] is empty for invalid molecules
+                } finally {
+                  mol?.delete();
+                }
               }
             }
             let rColName = '';
@@ -145,8 +155,37 @@ export function rGroupAnalysis(col: DG.Column): void {
             col.dataFrame.columns.add(rCol);
             latestAnalysisCols.push(rColName);
           }
-
+          //create column for highlight of uncommon structure
+          const highlightColName = `r-groups-highlight_${rGroupPrefixIdx}`;
+          let coreMol: RDMol | null = null;
+          let coreMolWithoutRGroups: RDMol | null = null;
+          try {
+            //const coreWithoutRGroups = core.replaceAll('R#', '*');
+            coreMol = module.get_mol(core);
+            const smiles = coreMol.get_smiles();
+            //remove r groups from smiles string to make highlight more precise
+            const newSmiles = smiles.replaceAll(/(\[\d+\*\])/g, '').replaceAll('()', '');
+            coreMolWithoutRGroups = module.get_mol(newSmiles);
+            const substructCol = DG.Column.fromType('object', highlightColName, col.dataFrame.rowCount)
+              .init((i) => getUncommonAtomsAndBonds(col.get(i), coreMolWithoutRGroups, module, '#bc131f', true));
+            col.dataFrame.columns.add(substructCol);
+            latestAnalysisCols.push(highlightColName);
+            col.temp[SUBSTRUCT_COL] = highlightColName;
+          } finally {
+            coreMol?.delete();
+            coreMolWithoutRGroups?.delete();
+          }
+          //create boolean column for match/non match
+          const isHitCol = DG.Column.bool(`${rGroupPrefixIdx ? `isHit_${rGroupPrefixIdx}` : `isHit`}`, res.rowCount)
+            .init((i) => unmatchedItems[i] !== res.columns.length - 1);
+          col.dataFrame.columns.add(isHitCol);
+          latestAnalysisCols.push(isHitCol.name);
+          //filter out unmatched values
+          const filterUnmatched = DG.BitSet.create(res.rowCount).init((i) => isHitCol.get(i));
+          col.dataFrame.filter.copyFrom(filterUnmatched);
           const view = grok.shell.getTableView(col.dataFrame.name);
+          //make highlight column invisible
+          view.grid.col(highlightColName)!.visible = false;
           if (visualAnalysisCheck.value && view) {
             latestTrellisPlot = view.trellisPlot({
               xColumnNames: [res.columns.byIndex(1).name], // column 0 is Core column
@@ -164,4 +203,10 @@ export function rGroupAnalysis(col: DG.Column): void {
     });
   dlg.show();
   dlg.initDefaultHistory();
+}
+
+function removeLatestAnalysis(col: DG.Column) {
+  latestTrellisPlot?.close();
+  latestAnalysisCols.forEach((colName: string) => col.dataFrame.columns.remove(colName));
+  latestAnalysisCols = [];
 }
